Move Report to a separate file
